Move check_unique_constraints function in constraints.py
